pyproject.toml
[tool.poetry]
name = "knowledge-base-search"
version = "1.0.0"
description = "AI-Powered Knowledge Base Search & Enrichment System"
authors = ["Hardik Khichi <hardikkhichi.iitd@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
PyPDF2 = "^3.0.1"
python-docx = "^1.1.0"
fpdf2 = "^2.7.0"
chromadb = "^0.4.18"
sentence-transformers = "^2.2.2"
openai = "^1.3.5"
aiofiles = "^23.2.1"
httpx = "^0.25.2"
tenacity = "^8.2.3"
tiktoken = "^0.5.2"
numpy = "^1.24.0,<2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
pre-commit = "^3.5.0"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers --disable-warnings"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

